Refactor tests with helper
